Reparo bug que rompia al crear un nuevo DosageUnit en addMedication Screen

+++ app/src/main/java/ar/ort/edu/proyecto_final_grupo_4/data/repository/MedicationRepositoryImpl.kt
@@ -31,7 +31,7 @@ class MedicationRepositoryImpl @Inject constructor(
         return medicationDao.getMedicationsByUser(userId)
     }
 
-    override suspend fun getById(id: Int): Medication? {
+    override suspend fun getById(id: Long): Medication? {
         return medicationDao.getById(id)
     }
